refactor(LegGenerator): test 코드에 존재하는 랜덤 방향 생성기 클래스 분리
