+++ src/test/java/FixedDirectionGenerator.java
@@ -1,0 +1,16 @@
+
+import java.util.List;
+
+public class FixedDirectionGenerator implements DirectionGenerator {
+    List<Direction> fixedList;
+    Integer index = 0;
+
+    FixedDirectionGenerator(List<Direction> fixedList) {
+        this.fixedList = fixedList;
+    }
+
+    @Override
+    public Direction generate() {
+        return fixedList.get(index++);
+    }
+}
